Shell script updated
Using cURL instead of wget, so we can catch it erroring out when it fails to fetch a list from the array and stop the script completely before the blocklist is written.
This way we prevent the script from potentially removing large chunks of the blocklist and we preserve yesterdays list.